add new function get_schoolkid commendation teacher is not random now, his id is from last lesson

+++ hack_the_diary.py
@@ -1,16 +1,22 @@
 import random
 
 from datacenter.models import Schoolkid, Mark, Chastisement, Teacher, Lesson, Commendation
-from django.core.exceptions import ObjectDoesNotExist, MultipleObjectsReturned
+
+
+def get_schoolkid(school_kid):
+    try:
+        schoolkid = Schoolkid.objects.get(full_name__contains=school_kid)
+        return schoolkid
+    except Schoolkid.DoesNotExist:
+        print('ФИО ученика не найдено, возможно у вас опечатка?')
+    except Schoolkid.MultipleObjectsReturned:
+        print('Уточните ФИО ученика, наёдено более одной записи')
 
 
 def replace_bad_marks(school_kid):
-    try:
-        schoolkid = Schoolkid.objects.get(full_name__contains=school_kid)
-    except ObjectDoesNotExist:
-        return 'Исправьте опечатку в имени ученика'
-    except MultipleObjectsReturned:
-        return 'Уточните ФИО ученика'
+    schoolkid = get_schoolkid(school_kid)
+    if not schoolkid:
+        return
     bad_marks = Mark.objects.filter(schoolkid_id=schoolkid.id, points__in=[2, 3])
     good_mark_points = random.randint(4, 5)
     for bad_mark in bad_marks:
@@ -20,12 +26,9 @@ def replace_bad_marks(school_kid):
 
 
 def delete_chastisements(school_kid):
-    try:
-        schoolkid = Schoolkid.objects.get(full_name__contains=school_kid)
-    except ObjectDoesNotExist:
-        return 'Исправьте опечатку в имени ученика'
-    except MultipleObjectsReturned:
-        return 'Уточните ФИО ученика'
+    schoolkid = get_schoolkid(school_kid)
+    if not schoolkid:
+        return
     chastisements = Chastisement.objects.filter(schoolkid_id=schoolkid.id)
     chastisements.delete()
     return chastisements.count()
@@ -38,21 +41,22 @@ def create_commendation(school_kid, subject_title):
     commendations = commendations.replace('\n', '')
     commendations = tuple(commendations.split(sep='!, '))
     random_commendation = random.choice(commendations)
+    schoolkid = get_schoolkid(school_kid)
+    if not schoolkid:
+        return
     try:
-        schoolkid = Schoolkid.objects.get(full_name__contains=school_kid)
-    except ObjectDoesNotExist:
-        return 'Исправьте опечатку в имени ученика'
-    except MultipleObjectsReturned:
-        return 'Уточните ФИО ученика'
-    last_lesson = Lesson.objects.filter(group_letter=schoolkid.group_letter,
-                                        year_of_study=schoolkid.year_of_study,
-                                        subject__title=subject_title).last()
-    if last_lesson is None:
-        return 'Исправьте опечатку в названии предмета'
-    random_teacher_id = random.choice(Teacher.objects.all()).id
-    Commendation.objects.create(text=random_commendation,
-                                schoolkid_id=schoolkid.id,
-                                created=last_lesson.date,
-                                subject_id=last_lesson.subject.id,
-                                teacher_id=random_teacher_id)
+        last_lesson = Lesson.objects.filter(group_letter=schoolkid.group_letter,
+                                            year_of_study=schoolkid.year_of_study,
+                                            subject__title=subject_title).last()
+        teacher_id = Teacher.objects.get(full_name=last_lesson.teacher).id
+
+        Commendation.objects.create(text=random_commendation,
+                                    schoolkid_id=schoolkid.id,
+                                    created=last_lesson.date,
+                                    subject_id=last_lesson.subject.id,
+                                    teacher_id=teacher_id)
+    except AttributeError:
+        return 'Уточните название предмета'
     return random_commendation
+
+
